requirements.txt
diffusers==0.24.0
faster_whisper==0.10.0
gTTS==2.4.0
openai==1.3.9
pinecone_client==2.2.4
pymongo==4.6.1
python_bcrypt==0.3.2
sentence_transformers==2.2.2
streamlit==1.29.0
torch==2.1.1
accelerate==0.25.0
bcrypt==4.1.1
